Abstract generic code for modelling debug info from DwarfSections into DebugInfoBase
Rename subclass DwarfSections to DwarfDebugInfo
